Separation SymbolTable and Activity visitors
